+++ src/main/java/com/aimprosoft/aimlearning/models/Employee.java
@@ -10,28 +10,35 @@ import java.util.Date;
 
 @With
 public class Employee {
+
     private int id;
+
     @NotEmpty(message = "Firstname can not be empty")
     @NotNull(message = "Firstname can not be null")
     @Length(max = 30, message = "Firstname can not be greater than 30")
     private String firstName;
+
     @NotEmpty(message = "Lastname can not be empty")
     @NotNull(message = "Lastname can not be null")
     @Length(max = 30, message = "Lastname can not be greater than 30")
     private String lastName;
+
     @NotEmpty(message = "email can not be empty")
     @NotNull(message = "email can not be null")
     @Length(max = 75, message = "Firstname can not be greater than 30")
     @CheckWith(value = IsUniqueEmail.class, message = "employee with such email exists")
     @Email(message = "wrong email format")
     private String email;
+
     @NotEmpty(message = "Salary can not be empty")
     @NotNull(message = "Salary can not be null")
     @Min(value = 1, message = "salary can not be smaller than 1")
-    private int salary;
+    private double salary;
+
     @NotEmpty(message = "Hire date can not be empty")
     @NotNull(message = "Hire date can not be null")
     private Date hireDate;
+
     @NotEmpty(message = "Id department can not be empty")
     @NotNull(message = "Id department can not be null")
     @Min(value = 1, message = "Id department can not be smaller than 1")
@@ -67,7 +74,7 @@ public class Employee {
         return lastName;
     }
 
-    public int getSalary() {
+    public double getSalary() {
         return salary;
     }
 
@@ -82,7 +89,7 @@ public class Employee {
     public Employee() {
     }
 
-    public Employee(String firstName, String lastName, String email, int salary, Date hireDate, int idDepartment) {
+    public Employee(String firstName, String lastName, String email, double salary, Date hireDate, int idDepartment) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
@@ -91,7 +98,7 @@ public class Employee {
         this.idDepartment = idDepartment;
     }
 
-    public Employee(int id, String firstName, String lastName, String email, int salary, Date hireDate, int idDepartment) {
+    public Employee(int id, String firstName, String lastName, String email, double salary, Date hireDate, int idDepartment) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
